Update mxsfb.c to current version
The lcdif driver mxsfb.c was still on an older version. Update to
u-boot-2021.04 but keep F&S extensions.

In arch/arm/include/asm/arch-imx8m/imx-regs.h, add parentheses around
MXS_LCDIF_BASE to fix usage.

Add const to first argument of vidconsole_putc() and vidconsole_puts().

Change isaBase in struct hraphic_device to unsigned long to handle 64-bit
pointers correctly.

+++ include/video_fb.h
@@ -32,7 +32,7 @@
 /******************************************************************************/
 
 typedef struct graphic_device {
-    unsigned int isaBase;
+    unsigned long isaBase;
     unsigned int pciBase;
     unsigned int dprBase;
     unsigned int vprBase;
